+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release notes
 
+### Pending
+
+- Added `allowUnpersistedOperation` option, allowing arbitrary operations to be issued under controlled circumstances
+
 ### v0.0.3
 
 - Relay compiler should now work with no additional config